huge fixes and improvements overall

+++ src/main/java/com/kingodogo/buildscape/item/MangroveBoatItem.java
@@ -56,7 +56,7 @@ public class MangroveBoatItem extends Item {
                 for (Entity entity : list) {
                     AABB aabb = entity
                             .getBoundingBox()
-                            .inflate((double) entity.getPickRadius());
+                            .inflate(entity.getPickRadius());
                     if (aabb.contains(vec31)) {
                         return InteractionResultHolder.pass(itemstack);
                     }
